+++ iniesta/sessions.py
@@ -1,4 +1,17 @@
 import aiobotocore
+from insanic.conf import settings
+from insanic.functional import empty
+
+class AWSCredentials(object):
+
+    def __init__(self, type):
+        self.value = empty
+        self.type = type
+
+    def __get__(self, instance, owner):
+        if self.value is empty:
+            self.value = getattr(settings, f'INIESTA_{self.type}', None) or getattr(settings, self.type, None)
+        return self.value
 
 
 class BotoSession:
@@ -9,3 +22,6 @@ class BotoSession:
         if cls.session is None:
             cls.session = aiobotocore.get_session(loop=loop)
         return cls.session
+
+    aws_access_key_id = AWSCredentials('AWS_ACCESS_KEY_ID')
+    aws_secret_access_key = AWSCredentials('AWS_SECRET_ACCESS_KEY')