+++ iniesta/sqs/client.py
@@ -83,8 +83,8 @@ class SQSClient:
         if queue_name not in cls.queue_urls:
             try:
                 async with session.create_client('sqs', endpoint_url=endpoint_url,
-                                                 aws_access_key_id=settings.INIESTA_AWS_ACCESS_KEY_ID,
-                                                 aws_secret_access_key=settings.INIESTA_AWS_SECRET_ACCESS_KEY) as client:
+                                                 aws_access_key_id=BotoSession.aws_access_key_id,
+                                                 aws_secret_access_key=BotoSession.aws_secret_access_key) as client:
                     response = await client.get_queue_url(QueueName=queue_name)
             except botocore.exceptions.ClientError as e:
                 error_message = f"[{e.response['Error']['Code']}]: {e.response['Error']['Message']} {queue_name}"
@@ -135,8 +135,8 @@ class SQSClient:
         session = BotoSession.get_session()
 
         async with session.create_client('sqs', endpoint_url=self.endpoint_url,
-                                         aws_access_key_id=settings.INIESTA_AWS_ACCESS_KEY_ID,
-                                         aws_secret_access_key=settings.INIESTA_AWS_SECRET_ACCESS_KEY) as client:
+                                         aws_access_key_id=BotoSession.aws_access_key_id,
+                                         aws_secret_access_key=BotoSession.aws_secret_access_key) as client:
             policy_attributes = await client.get_queue_attributes(
                 QueueUrl=self.queue_url,
                 AttributeNames=['Policy']
@@ -268,8 +268,8 @@ class SQSClient:
 
         session = BotoSession.get_session()
         client = session.create_client('sqs', endpoint_url=self.endpoint_url,
-                                       aws_access_key_id=settings.INIESTA_AWS_ACCESS_KEY_ID,
-                                       aws_secret_access_key=settings.INIESTA_AWS_SECRET_ACCESS_KEY)
+                                       aws_access_key_id=BotoSession.aws_access_key_id,
+                                       aws_secret_access_key=BotoSession.aws_secret_access_key)
 
         try:
             while self._loop.is_running() and self._receive_messages:
@@ -370,9 +370,3 @@ class SQSClient:
         :rtype: SQSMessage
         """
         return SQSMessage(self, message)
-
-
-
-
-
-