+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 - CHORE: MO documentations!!!
 - CHORE: change extra requires from ``deploy`` to ``cli``
 - CHORE: remove used parameters
+- FEAT: allow aws credentials to be loaded from iniesta then fallback to insanic